+++ midban/src/main/java/com/cafedered/midban/view/fragments/OrderNewDispositionFragment.java
@@ -1120,7 +1120,7 @@ y por tanto al escribir un valor en el commercial_partner_id se propaga a las di
             if (lines.get(i).getName() != null) {
                 lineValues.put("description", lines.get(i).getName());
             }
-            lineValues.put("product_uom_qty", lines.get(i).getProductUomQuantity().intValue());
+            lineValues.put("product_uom_qty", lines.get(i).getProductUomQuantity().doubleValue());
             lineValues.put("product_uos_qty", lines.get(i).getProductUosQuantity().doubleValue());
             Product product = ProductRepository.getInstance().getById(lines.get(i).getProductId().longValue());
             if (product != null) {
